refactor(chat): 채팅방 컴포넌트 파일 구조 수정(#31)
